fix: defaultArgs should work for all commands
This is how we check if the defaultArg is in the command already: `if (!defaultArgs.hasOwnProperty(arg) || command.match(arg)) continue;`.

If the command string looks something like this: `/circleci/project/node_modules/.bin/tsc ..bla`, then args such as the default project arg would not get added because `project` is already in the command string. This is causing a variety of issues. To solve this, I simply made the arg keys more specific.